add UT for refactored metrics
